refactor: migrate Java SDK references to Kotlin SDK and update package structure
